Title: 세 수의 합과 평균 구하기, Time: 194ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240219/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -5,10 +5,10 @@ import java.util.Scanner;
 public class Main {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
-    float a = sc.nextFloat();
-    float b = sc.nextFloat();
-    float c = sc.nextFloat();
-    System.out.printf("%.0f\n", a + b + c);
+    int a = sc.nextInt();
+    int b = sc.nextInt();
+    int c = sc.nextInt();
+    System.out.printf("%d\n", a + b + c);
     System.out.printf("%d", (int) (a + b + c) / 3);
     // 여기에 코드를 작성해주세요.
   }
